Restructure file organization for dashboard
