+++ lib/controllers/ephemeral/dashboard/account_dashboard_controller.dart
@@ -7,30 +7,34 @@ import 'package:pispapp/repositories/interfaces/i_transaction_repository.dart';
 
 class AccountDashboardController extends GetxController {
   AccountDashboardController(this._transactionRepo);
+
   Account selectedAccount;
+
   List<Transaction> transactionList = <Transaction>[];
+
   bool isLoading = true;
 
   ITransactionRepository _transactionRepo;
+
   bool noAccounts = true;
 
-  Future<void> onRefresh() async {
+  Future<void> refresh() async {
     isLoading = true;
     update();
+
     await getLinkedAccounts();
+
     if (Get.find<AccountController>().accounts.isEmpty) {
       noAccounts = true;
     } else {
       noAccounts = false;
-      await setSelectedAccount(Get.find<AccountController>().accounts.elementAt(0));
+      await setSelectedAccount(
+        Get.find<AccountController>().accounts.elementAt(0),
+      );
     }
     isLoading = false;
 
     update();
-  }
-
-  void onAccountTileTap(Account acc) {
-    setSelectedAccount(acc);
   }
 
   Future<void> setSelectedAccount(Account acc) async {
@@ -39,13 +43,12 @@ class AccountDashboardController extends GetxController {
   }
 
   Future<void> updateTransactions() async {
-    final userId = Get.find<AuthController>().user.uid;
-    transactionList = await _transactionRepo.getTransactions(
-        userId, selectedAccount.sourceAccountId);
+    final userId = Get.find<AuthController>().user.id;
+    transactionList = await _transactionRepo.getTransactions(userId);
     update();
   }
 
   Future<void> getLinkedAccounts() async {
-    await Get.find<AccountController>().getAllLinkedAccounts();
+    await Get.find<AccountController>().getLinkedAccounts();
   }
 }
