+++ lib/ui/pages/dashboard.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:pispapp/controllers/ephemeral/dashboard_controller.dart';
+import 'package:pispapp/controllers/ephemeral/dashboard/dashboard_controller.dart';
 import 'package:pispapp/ui/theme/light_theme.dart';
 
 class Dashboard extends StatelessWidget {
